moved /screens files to /relay
